+++ openedx/features/ondemand_email_preferences/utils.py
@@ -1,0 +1,5 @@
+from datetime import timedelta
+
+
+def get_next_date(today, module_date):
+    return str(today + timedelta(days=module_date))
